moving common classes to common
instead of being in wear or mobile they can be shared by both targets
