+++ common/src/main/java/com/github/qjcg/soundapp/common/AudioService.java
@@ -1,4 +1,4 @@
-package com.github.qjcg.soundapp;
+package com.github.qjcg.soundapp.common;
 
 import android.app.Service;
 import android.content.BroadcastReceiver;
