+++ mobile/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <service android:name="com.github.qjcp.soundapp.common.SoundFilterIntentService" />
+        <service android:name="com.github.qjcg.soundapp.common.SoundFilterIntentService" />
 
         <meta-data android:name="com.google.android.gms.version" android:value="@integer/google_play_services_version" />
     </application>
